Create user

+++ pages/sistema/users.js
@@ -1,15 +1,20 @@
-import { useEffect, useState } from "react"
-import useSound from 'use-sound'
-import { Alert_path } from "../../../components/alert/alert"
-import { Fetch_Data } from "../../../components/fetch/fetch_data"
-import { Fetch_Next } from "../../../components/fetch/fetch_next"
-import { Attributes } from "../../../components/helpers/consts"
-import { User } from "../../../components/helpers/database"
-import Layout from "../../../components/Layout"
-import Table_Body from "../../../components/table/tb_body"
-import { Text } from "../../../components/text/text"
+import { useRouter } from "next/router";
+import { useEffect, useState } from "react";
+import useSound from 'use-sound';
+import { Alert_path, Alert_show } from "../../components/alert/alert";
+import Btn_Standard from "../../components/buttons/btn_standard";
+import { Fetch_Data } from "../../components/fetch/fetch_data";
+import { Fetch_Next } from "../../components/fetch/fetch_next";
+import { Attributes } from "../../components/helpers/consts";
+import { User } from "../../components/helpers/database";
+import Layout from "../../components/layout";
+import Table_Body from "../../components/table/tb_body";
+import { Errors } from "../../components/text/error";
+import { Text } from "../../components/text/text";
 
 const Users = () => {
+    // Get the router
+    const router = new useRouter();
     // Save the information
     const [data, setData] = useState([]);
     // Empty data (dont reload)
@@ -36,6 +41,10 @@ const Users = () => {
         Fetch_Next(User.putUser(newRow[Attributes.Username]), play, fetchData, "put", { data: newRow })
     }
 
+    function changePage() {
+        router.push('/signup')
+    }
+
     // Selected headers (filtered by name)
     const headers = [
         "Username", "Name", "Admin"
@@ -51,10 +60,18 @@ const Users = () => {
         "Refrescar": fetchData,
         "editable": {
             onRowDelete: (row) => new Promise((res, rej) => {
+                if (row.Username == "admin") {
+                    Alert_show(Errors.Protected, play)
+                    return res()
+                }
                 deleteMaterial(row)
                 res([])
             }),
             onRowUpdate: (newRow, oldRow) => new Promise((res, rej) => {
+                if (newRow.Username == "admin") {
+                    Alert_show(Errors.Protected, play)
+                    return res()
+                }
                 updateMaterial(newRow, oldRow)
                 res([])
             }),
@@ -66,6 +83,9 @@ const Users = () => {
             <div className="my-3">
                 <h1 className="text-2xl text-gray-800 font-light">{Text.Users}</h1>
                 <div className='my-3'>
+                    <div className="flex flex-row-reverse">
+                        <Btn_Standard btnName={Text.Create} btnClick={changePage} />
+                    </div>
                     <Table_Body
                         tableTitle={Text.Users}
                         tableData={data}
